+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 
 ### Snowpark Local Testing Updates
 
+- Added local testing support for reading files with `SnowflakeFile` using local file paths.
+
 #### Bug Fixes
 
 - Fixed a bug when processing windowed functions that lead to incorrect indexing in results.